[Fix] Jwt token 만료시간 타입 long으로 변경

+++ back/src/main/java/travelRepo/global/security/jwt/JwtProcessor.java
@@ -21,7 +21,7 @@ public class JwtProcessor {
     String secretKey;
 
     @Value("${jwt.expiration}")
-    int expiration;
+    long expiration;
 
     @Value("${jwt.prefix}")
     String prefix;
